usb: wMaxPacketSize is 16-bit
In USB endpoint descriptors wMaxPacketSize is the following 16-bit
little-endian bit field.

  wMaxPacketSize[0:10]   maximum size of data payload
  wMaxPacketSize[11:12]  number of transactions per microframe
  wMaxPacketSize[13:15]  reserved MBZ

Fixes #5624

+++ repos/os/src/lib/genode_c_api/usb_client.cc
@@ -29,15 +29,15 @@ struct Device;
 
 struct Endpoint : List_model<Endpoint>::Element
 {
-	uint8_t const address;
-	uint8_t const attributes;
-	uint8_t const max_packet_size;
+	uint8_t  const address;
+	uint8_t  const attributes;
+	uint16_t const max_packet_size;
 
 	Endpoint(Node const &n)
 	:
 		address(n.attribute_value<uint8_t>("address", 0xff)),
 		attributes(n.attribute_value<uint8_t>("attributes", 0xff)),
-		max_packet_size(n.attribute_value<uint8_t>("max_packet_size", 0)) {}
+		max_packet_size(n.attribute_value<uint16_t>("max_packet_size", 0)) {}
 
 	bool matches(Node const &node) const {
 		return address == node.attribute_value<uint8_t>("address", 0xff); }
